Graficas Practica4

+++ PrimerParcial/Practica4/fibonacci-r.c
@@ -1,7 +1,7 @@
 #include <stdio.h>
 #include <time.h>
 
-unsigned long long int Fibonacci(unsigned long long int num){
+unsigned int Fibonacci(unsigned int num){
     if (num == 0)
         return 0;
     else if (num == 1)
@@ -11,14 +11,14 @@ unsigned long long int Fibonacci(unsigned long long int num){
 }
 
 int main() {
-    unsigned long long int num;
+    unsigned int num;
     printf("Escribe el numero del fibonacci que deseas obtener ");
-    scanf("%llu", &num);
+    scanf("%d", &num);
     clock_t start, end;
     start = clock();
-    unsigned long long int fib = Fibonacci(num);
+    unsigned int fib = Fibonacci(num);
     end = clock();
-    printf("\nFibonacci de %llu es %llu", num, fib);
+    printf("\nFibonacci de %d es %d", num, fib);
     double time_taken = ((double)end - start) / CLOCKS_PER_SEC;
     printf("\nTiempo tomado fue: %.10f segundos\n", time_taken);
     return 0;
